fix(verify): ensure single POST /api/certified/verify handler; add diagnostics; enforce CORS; keep contract
