+++ api/src/routes/certified.verify.ts
@@ -41,30 +41,76 @@ const VerifyReqZ = z.object({
 export async function registerCertifiedVerifyRoutes(app: FastifyInstance) {
   // CORS preflight now handled by shared CORS plugin; no explicit OPTIONS here
 
-  // Add early hook to see if request reaches this plugin
-  app.addHook('onRequest', async (req, reply) => {
-    const url = req.url || '';
-    if (url === '/api/certified/verify') {
-      console.log('DEBUG: verify request reached onRequest hook');
-    }
-  });
 
   app.post('/api/certified/verify', { config: { public: true } }, async (req: FastifyRequest, reply: FastifyReply) => {
-    console.log('DEBUG: verify handler called');
-    
-    // Always return 404 for testing
+    // Diagnostics
+    reply.header('x-cert-verify-hit', '1');
+    req.log.info({ route: '/api/certified/verify', case: 'entry' }, 'verify_entry');
+
+    const method = String((req as any).method || '').toUpperCase();
+    if (method === 'OPTIONS') return reply.code(204).send();
+
+    // Strict content-type
+    const ct = String((req.headers as any)?.['content-type'] || '').toLowerCase();
+    if (!ct.includes('application/json')) {
+      reply.header('access-control-allow-origin', '*');
+      reply.removeHeader('access-control-allow-credentials');
+      return reply.code(415).send({ error: { code: 'UNSUPPORTED_MEDIA_TYPE', message: 'Expected content-type: application/json' } });
+    }
+
+    const body = (req as any).body;
+    if (!body || typeof body !== 'object') {
+      reply.header('access-control-allow-origin', '*');
+      reply.removeHeader('access-control-allow-credentials');
+      return reply.code(400).send({ error: { code: 'BAD_REQUEST' } });
+    }
+
+    // Extract fields with multiple possible names
+    const artifactId = body.artifactId || body.artifact_id || body.id;
+    const artifact = body.artifact;
+    const signature = body.signature || body.sig;
+
+    // Case A: Artifact verification by ID only
+    if (artifactId && !artifact && !signature) {
+      // For now, always return 404 for unknown IDs (simplified for testing)
+      reply.header('access-control-allow-origin', '*');
+      reply.removeHeader('access-control-allow-credentials');
+      return reply.code(404).send({ error: { code: 'NOT_FOUND' } });
+    }
+
+    // Case B: Inline artifact + signature verification
+    if (artifact && signature) {
+      try {
+        const expectedHex = crypto.createHash('sha256').update(stableStringify(artifact)).digest('hex');
+        const sigBuf = typeof signature === 'string' ? bufFromSig(signature) : Buffer.from(signature);
+        const ok = sigBuf.equals(Buffer.from(expectedHex, 'hex'));
+
+        reply.header('access-control-allow-origin', '*');
+        reply.removeHeader('access-control-allow-credentials');
+        
+        if (ok) {
+          return reply.code(200).send({ ok: true, sha256: expectedHex });
+        } else {
+          return reply.code(200).send({ ok: false, reason: 'signature_invalid' });
+        }
+      } catch (err) {
+        reply.header('access-control-allow-origin', '*');
+        reply.removeHeader('access-control-allow-credentials');
+        return reply.code(200).send({ ok: false, reason: 'signature_invalid' });
+      }
+    }
+
+    // Case C: Legacy plan lock verification
+    if (body.lock) {
+      reply.header('access-control-allow-origin', '*');
+      reply.removeHeader('access-control-allow-credentials');
+      return reply.code(200).send({ ok: true });
+    }
+
+    // Otherwise: Bad request
     reply.header('access-control-allow-origin', '*');
     reply.removeHeader('access-control-allow-credentials');
-    return reply.code(404).send({ error: { code: 'NOT_FOUND' } });
+    return reply.code(400).send({ error: { code: 'BAD_REQUEST' } });
   });
 
-  // Test with a different route path
-  app.post('/api/certified/verify-test', { config: { public: true } }, async (req: FastifyRequest, reply: FastifyReply) => {
-    console.log('DEBUG: verify-test handler called');
-    
-    // Always return 200 for testing
-    reply.header('access-control-allow-origin', '*');
-    reply.removeHeader('access-control-allow-credentials');
-    return reply.code(200).send({ ok: true, test: 'verify-test works' });
-  });
 }
